Unignore .idea/ directory in .gitignore
The .idea/ directory was previously ignored, but has now been unignored. This directory contains project-specific settings files and configurations that are essential to maintain consistency when the project is cloned or pulled to different machines or by different developers. This .idea/ directory is vital and should be tracked by Git. Thus, the changes to unignore it.