Refactor database session management
In SQLAlchemy __init__.py and session.py, the naming and structure of classes for retrieving database sessions have been refactored. 'GetSession' is now 'DatabaseSessionFactory' and 'GetDB' is now 'DatabaseSessionManager'. These changes improve readability and better represent their functionality. This is part of the ongoing effort to improve code base maintainability and clarity.