Update library version and add pagination dependency
Upgraded the library 'fastutils-hmarcuzzo' from version '0.3.24' to '0.4.22'. This includes changes to support advance pagination features. Additionally, added the 'fastapi-pagination' package to dependencies registered in 'pyproject.toml' and 'poetry.lock'.

The 'fastapi-pagination' package is added to utilize its dynamic sorting and searching capabilities on multiple fields. This would enhance the PaginationUtils class to handle complicated scenarios. The update not only improves the usability of PaginationUtils but also increases application efficiency by leveraging the database's power for sorting and searching.